# public/postman/schemas/api.yml
openapi: 3.0.3
info:
  title: Arcane City
  version: 2025.06.09
  description: REST API for working with Arcane City data
  contact:
    name: Geoff Maddock
    email: geoff.maddock@gmail.com
servers:
  - url: https://arcane.city
    description: Production
  - url: https://stage.arcane.city
    description: Stage
  - url: https://dev.arcane.city
    description: Development
...
components:
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  schemas:
...
    Thread:
      type: object
      required:
        - name
      properties:
        id:
          type: integer
          readOnly: true
          example: 1
        name:
          type: string
          maxLength: 255
          description: A name of a thread
          example: Forum
        slug:
          type: string
          maxLength: 255
          description: A unique identifier name for the thread in kebab-case
          example: thread-slug
        description:
          type: string
          example: An extensive breakdown of the thread
          description: An extensive breakdown of the thread
        visibility_id:
          type: integer
          readOnly: true
          example: 1
          description: Relation to the visibility table that defines the visibility of the thread
        forum_id:
          type: integer
          readOnly: true
          example: 1
          description: Relation to the forum table that defines the forum of the thread
        created_at:
          type: string
          example: "2018-03-20T09:12:28Z"
          format: date-time
          description: Date and time that the entity type was created
          readOnly: true
        updated_at:
          type: string
          example: "2018-03-20T09:12:28Z"
          format: date-time
          description: Date and time that the entity type was last updated
          readOnly: true
    Threads:
      allOf: [$ref: "#/components/schemas/Pagination"]
      type: object
      properties:
        data:
          type: array
          description: List of the current page of threads
          items:
            "$ref": "#/components/schemas/Thread"
...
    Pagination:
      type: object
      properties:
        current_page:
          description: The current page in the paginated list
          type: integer
        first_page_url:
          description: URL for the first page
          type: string
          example: "https://dev.arcane.city/api/resources?page=1"
        from:
          description: First element in the current page
          type: integer
          example: 1
        last_page:
          description: Last page of results
          type: integer
          example: 1
        last_page_url:
          description: URL for the first page
          type: string
          example: "https://dev.arcane.city/api/resources?page=100"
        links:
          description: Array of links used to paginate the list of entities
          type: array
          items: {}
        next_page_url:
          description: URL for the next page
          type: string
          example: "https://dev.arcane.city/api/resources?page=2"
        path:
          description: Path for the primary route
          type: string
          example: "https://dev.arcane.city/api/resources"
        per_page:
          description: Number of results per page
          type: integer
          example: 25
        prev_page_url:
          description: URL for the previous page
          type: string
          example: "https://dev.arcane.city/api/resources?page=0"
        to:
          description: Last element in the current page
          type: integer
          example: 25
        total:
          description: Total number of results from the request
          type: integer
          example: 100
        data:
          description: List of the current page of the paginated entity
          type: array
          items: {}
security:
  - basicAuth: []
paths:
...
  /api/threads:
    get:
      tags:
        - threads
      summary: Get Threads
      operationId: getThreads
      parameters:
        - name: filters[name]
          in: query
          required: false
          description: A filter query of the thread name
          schema:
            type: string
            example: Thread name
        - name: limit
          in: query
          required: false
          description: The limit on the number of results to return per page
          schema:
            type: integer
            example: 25
        - name: page
          in: query
          required: false
          description: The page number to return
          schema:
            type: integer
            example: 1
        - name: sort
          in: query
          required: false
          description: A column to be used to sort the query
          schema:
            type: string
            example: name
        - name: direction
          in: query
          required: false
          description: A string indicating the sort direction of the query (asc or desc)
          schema:
            type: string
            example: asc
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Threads"
    post:
      tags:
        - threads
      summary: Create thread
      operationId: createThread
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Thread"
      responses:
        "201":
          description: Successful response
          content:
            application/json: {}
  /api/threads/{threadId}:
    get:
      parameters:
        - name: threadId
          description: The unique identifier of the thread
          in: path
          required: true
          schema:
            description: The unique identifier of a thread
            type: integer
            readOnly: true
            example: 1
      tags:
        - threads
      summary: Get Thread
      operationId: getThreadId
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Thread"
    put:
      parameters:
        - name: threadId
          description: The unique identifier of the thread
          in: path
          required: true
          schema:
            description: The unique identifier of a thread
            type: integer
            readOnly: true
            example: 1
      tags:
        - threads
      summary: Update Thread (PUT — full replace)
      description: |
        Full replacement of the resource per REST PUT semantics. Optional scalar fields omitted from the body are reset to null, and any *_list relation arrays sync to exactly the supplied values — omitting a list detaches all rows for that relation. Use PATCH for partial updates.
      operationId: updateThreadById
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Thread"
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Thread"
    patch:
      parameters:
        - name: threadId
          description: The unique identifier of the thread
          in: path
          required: true
          schema:
            description: The unique identifier of a thread
            type: integer
            readOnly: true
            example: 1
      tags:
        - threads
      summary: Patch Thread (PATCH — partial update)
      description: |
        Partial update per REST PATCH semantics. Only the fields present in the body are modified; scalars and relations not included are left untouched. Validation rules from the corresponding *Request schema are applied as `sometimes` rules — if a field is supplied it must satisfy its constraints, but missing fields are not rejected. To clear a relation explicitly, send an empty array (e.g. `"tag_list": []`).
        Full replacement of the resource per REST PUT semantics. Optional scalar fields omitted from the body are reset to null, and any *_list relation arrays sync to exactly the supplied values — omitting a list detaches all rows for that relation. Use PATCH for partial updates.
      operationId: patchThreadById
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Thread"
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Thread"
    delete:
      parameters:
        - name: threadId
          description: The unique identifier of the thread
          in: path
          required: true
          schema:
            description: The unique identifier of a thread
            type: integer
            readOnly: true
            example: 1
      tags:
        - threads
      summary: Delete Thread
      operationId: deleteThreadById
      responses:
        "204":
          description: Successful response
          content:
            application/json: {}
...